create reusable components and modify descriptions
